[Fix] MongoDB 필드명 변경&카테고리 추가

+++ src/main/java/com/click/batchServer/domain/entity/Category.java
@@ -22,7 +22,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "CATEGORY_ID")
-    private Long categoryId;
+    private Integer categoryId;
 
     @Column(name = "CATEGORY_NAME")
     private String categoryName;
